+++ test/utils/compatibility-test.js
@@ -12,6 +12,7 @@ vows.describe(Compatibility)
         assert.isTrue(options.colors.opacity);
         assert.isTrue(options.properties.colors);
         assert.isFalse(options.properties.backgroundClipMerging);
+        assert.isFalse(options.properties.backgroundOriginMerging);
         assert.isFalse(options.properties.backgroundSizeMerging);
         assert.isFalse(options.properties.iePrefixHack);
         assert.isFalse(options.properties.ieSuffixHack);
@@ -46,6 +47,7 @@ vows.describe(Compatibility)
       'gets merged options': function(options) {
         assert.isTrue(options.colors.opacity);
         assert.isFalse(options.properties.backgroundClipMerging);
+        assert.isFalse(options.properties.backgroundOriginMerging);
         assert.isFalse(options.properties.backgroundSizeMerging);
         assert.isTrue(options.properties.colors);
         assert.isFalse(options.properties.iePrefixHack);
@@ -74,6 +76,7 @@ vows.describe(Compatibility)
       'gets template options': function(options) {
         assert.isFalse(options.colors.opacity);
         assert.isFalse(options.properties.backgroundClipMerging);
+        assert.isFalse(options.properties.backgroundOriginMerging);
         assert.isFalse(options.properties.backgroundSizeMerging);
         assert.isTrue(options.properties.colors);
         assert.isTrue(options.properties.iePrefixHack);
@@ -101,6 +104,7 @@ vows.describe(Compatibility)
       'gets template options': function(options) {
         assert.isFalse(options.colors.opacity);
         assert.isFalse(options.properties.backgroundClipMerging);
+        assert.isFalse(options.properties.backgroundOriginMerging);
         assert.isFalse(options.properties.backgroundSizeMerging);
         assert.isTrue(options.properties.colors);
         assert.isTrue(options.properties.iePrefixHack);
@@ -138,6 +142,7 @@ vows.describe(Compatibility)
       'gets calculated options': function(options) {
         assert.isTrue(options.colors.opacity);
         assert.isFalse(options.properties.backgroundClipMerging);
+        assert.isFalse(options.properties.backgroundOriginMerging);
         assert.isFalse(options.properties.backgroundSizeMerging);
         assert.isTrue(options.properties.colors);
         assert.isFalse(options.properties.iePrefixHack);
@@ -166,6 +171,7 @@ vows.describe(Compatibility)
         assert.isTrue(options.colors.opacity);
         assert.isTrue(options.properties.colors);
         assert.isFalse(options.properties.backgroundClipMerging);
+        assert.isFalse(options.properties.backgroundOriginMerging);
         assert.isFalse(options.properties.backgroundSizeMerging);
         assert.isTrue(options.properties.iePrefixHack);
         assert.isFalse(options.properties.ieSuffixHack);
@@ -193,6 +199,7 @@ vows.describe(Compatibility)
         assert.isTrue(options.colors.opacity);
         assert.isTrue(options.properties.colors);
         assert.isFalse(options.properties.backgroundClipMerging);
+        assert.isFalse(options.properties.backgroundOriginMerging);
         assert.isFalse(options.properties.backgroundSizeMerging);
         assert.isTrue(options.properties.iePrefixHack);
         assert.isFalse(options.properties.ieSuffixHack);